+++ notify_pod_client/lib/src/protocol/devices/devices.dart
@@ -18,7 +18,7 @@ abstract class Device implements _i1.SerializableModel {
     this.id,
     required this.userId,
     required this.idDevice,
-    required this.type,
+    this.type,
     required this.tokenFCM,
     this.notificationsLogs,
     required this.createdAt,
@@ -29,7 +29,7 @@ abstract class Device implements _i1.SerializableModel {
     int? id,
     required String userId,
     required String idDevice,
-    required _i2.DevicesType type,
+    _i2.DevicesType? type,
     required String tokenFCM,
     List<_i3.NotificationsLogs>? notificationsLogs,
     required DateTime createdAt,
@@ -41,7 +41,9 @@ abstract class Device implements _i1.SerializableModel {
       id: jsonSerialization['id'] as int?,
       userId: jsonSerialization['userId'] as String,
       idDevice: jsonSerialization['idDevice'] as String,
-      type: _i2.DevicesType.fromJson((jsonSerialization['type'] as int)),
+      type: jsonSerialization['type'] == null
+          ? null
+          : _i2.DevicesType.fromJson((jsonSerialization['type'] as int)),
       tokenFCM: jsonSerialization['tokenFCM'] as String,
       notificationsLogs: (jsonSerialization['notificationsLogs'] as List?)
           ?.map((e) =>
@@ -63,7 +65,7 @@ abstract class Device implements _i1.SerializableModel {
 
   String idDevice;
 
-  _i2.DevicesType type;
+  _i2.DevicesType? type;
 
   String tokenFCM;
 
@@ -89,7 +91,7 @@ abstract class Device implements _i1.SerializableModel {
       if (id != null) 'id': id,
       'userId': userId,
       'idDevice': idDevice,
-      'type': type.toJson(),
+      if (type != null) 'type': type?.toJson(),
       'tokenFCM': tokenFCM,
       if (notificationsLogs != null)
         'notificationsLogs':
@@ -112,7 +114,7 @@ class _DeviceImpl extends Device {
     int? id,
     required String userId,
     required String idDevice,
-    required _i2.DevicesType type,
+    _i2.DevicesType? type,
     required String tokenFCM,
     List<_i3.NotificationsLogs>? notificationsLogs,
     required DateTime createdAt,
@@ -133,7 +135,7 @@ class _DeviceImpl extends Device {
     Object? id = _Undefined,
     String? userId,
     String? idDevice,
-    _i2.DevicesType? type,
+    Object? type = _Undefined,
     String? tokenFCM,
     Object? notificationsLogs = _Undefined,
     DateTime? createdAt,
@@ -143,7 +145,7 @@ class _DeviceImpl extends Device {
       id: id is int? ? id : this.id,
       userId: userId ?? this.userId,
       idDevice: idDevice ?? this.idDevice,
-      type: type ?? this.type,
+      type: type is _i2.DevicesType? ? type : this.type,
       tokenFCM: tokenFCM ?? this.tokenFCM,
       notificationsLogs: notificationsLogs is List<_i3.NotificationsLogs>?
           ? notificationsLogs