+++ notify_pod_server/CHANGELOG.md
@@ -1,3 +1,11 @@
+# Changelog Notifypod_client
+
 ## 1.0.0
 
 - Initial version, created by Chivo
+
+## 1.0.1
+
+- Removed the deviceType enum.
+- Determining the device type is no longer required.
+- The device type is now automatically detected.